+++ src/typed_settings/cls_utils.py
@@ -88,6 +88,7 @@ class Attrs:
                 else:
                     is_nothing = field.default is attrs.NOTHING
                     is_factory = isinstance(field.default, cast(type, attrs.Factory))
+                    metadata = _get_metadata(field.metadata.get(constants.METADATA_KEY))
                     oinfo = types.OptionInfo(
                         parent_cls=r_cls,
                         path=f"{prefix}{field.name}",
@@ -103,7 +104,7 @@ class Attrs:
                         has_no_default=is_nothing,
                         default_is_factory=is_factory,
                         converter=field.converter,
-                        metadata=field.metadata.get(constants.METADATA_KEY, {}),
+                        metadata=metadata,
                     )
                     result.append(oinfo)
 
@@ -151,6 +152,7 @@ class Dataclasses:
                     is_factory = (
                         is_nothing and field.default_factory is not dataclasses.MISSING
                     )
+                    metadata = _get_metadata(field.metadata.get(constants.METADATA_KEY))
                     oinfo = types.OptionInfo(
                         parent_cls=r_cls,
                         path=f"{prefix}{field.name}",
@@ -166,7 +168,7 @@ class Dataclasses:
                         has_no_default=is_nothing,
                         default_is_factory=is_factory,
                         converter=None,
-                        metadata=field.metadata.get(constants.METADATA_KEY, {}),
+                        metadata=metadata,
                     )
                     result.append(oinfo)
 
@@ -271,7 +273,9 @@ class Pydantic:
                     iter_attribs(field.annotation, f"{prefix}{name}.")
                 else:
                     json_schema_extra = field.json_schema_extra or {}
-                    metadata = json_schema_extra.get("metadata", {})
+                    metadata_or_none = json_schema_extra.get(constants.METADATA_KEY, {})
+                    metadata = _get_metadata(metadata_or_none, field.description)
+
                     oinfo = types.OptionInfo(
                         parent_cls=r_cls,
                         path=f"{prefix}{name}",
@@ -293,7 +297,7 @@ class Pydantic:
                         has_no_default=field.is_required(),
                         default_is_factory=False,
                         converter=None,
-                        metadata=metadata.get(constants.METADATA_KEY, {}),
+                        metadata=metadata,
                     )
                     result.append(oinfo)
 
@@ -408,3 +412,28 @@ def group_options(
     grouper = groupby(options, key=keyfn)
     grouped_options = [(g_cls[1], tuple(g_opts)) for g_cls, g_opts in grouper]
     return grouped_options
+
+
+def _get_metadata(metadata_or_none: Any, default_help: Optional[str] = None) -> dict:
+    metadata = metadata_or_none if isinstance(metadata_or_none, dict) else {}
+
+    cli_defaults: dict[str, Any] = {}
+    if default_help:
+        cli_defaults["help"] = default_help
+    if "help" in metadata:
+        cli_defaults["help"] = metadata["help"]
+
+    click_config = {
+        **cli_defaults,
+        **metadata.get(constants.CLICK_METADATA_KEY, {}),
+    }
+    argparse_config = {
+        **cli_defaults,
+        **metadata.get(constants.ARGPARSE_METADATA_KEY, {}),
+    }
+    if click_config:
+        metadata[constants.CLICK_METADATA_KEY] = click_config
+    if argparse_config:
+        metadata[constants.ARGPARSE_METADATA_KEY] = argparse_config
+
+    return metadata