+++ tests/test_cls_utils.py
@@ -245,11 +245,26 @@ class TestAttrs:
 
         @attrs.define
         class GrandChild:
-            x: Optional[int] = attrs.field(factory=factory_fn)
+            x: Optional[int] = attrs.field(
+                factory=factory_fn,
+                metadata={
+                    "typed-settings": {
+                        "help": "grand child x",
+                        "argparse": {"metavar": "GX"},
+                        "click": {"metavar": "GX"},
+                    },
+                },
+            )
 
         @attrs.define
         class Child:
-            x: "float"  # Test resolving types
+            x: "float" = attrs.field(  # Test resolving types
+                metadata={
+                    "typed-settings": {
+                        "click": {"help": "child x", "metavar": "X"},
+                    },
+                },
+            )
             y: GrandChild
 
         @attrs.define
@@ -275,6 +290,9 @@ class TestAttrs:
                 default=attrs.NOTHING,
                 has_no_default=True,
                 default_is_factory=False,
+                metadata={
+                    "click": {"help": "child x", "metavar": "X"},
+                },
             ),
             types.OptionInfo(
                 parent_cls=GrandChild,
@@ -283,6 +301,11 @@ class TestAttrs:
                 default=attrs.Factory(factory_fn),
                 has_no_default=False,
                 default_is_factory=True,
+                metadata={
+                    "argparse": {"help": "grand child x", "metavar": "GX"},
+                    "click": {"help": "grand child x", "metavar": "GX"},
+                    "help": "grand child x",
+                },
             ),
             types.OptionInfo(
                 parent_cls=Parent,
@@ -424,11 +447,26 @@ class TestDataclasses:
 
         @dataclasses.dataclass
         class GrandChild:
-            x: Optional[int] = None
+            x: Optional[int] = dataclasses.field(
+                default=None,
+                metadata={
+                    "typed-settings": {
+                        "help": "grand child x",
+                        "argparse": {"metavar": "GX"},
+                        "click": {"metavar": "GX"},
+                    },
+                },
+            )
 
         @dataclasses.dataclass
         class Child:
-            x: "float"  # Test resolving types
+            x: "float" = dataclasses.field(  # Test resolving types
+                metadata={
+                    "typed-settings": {
+                        "click": {"help": "child x", "metavar": "X"},
+                    },
+                },
+            )
             y: GrandChild
 
         @dataclasses.dataclass
@@ -454,6 +492,9 @@ class TestDataclasses:
                 default=dataclasses.MISSING,
                 has_no_default=True,
                 default_is_factory=False,
+                metadata={
+                    "click": {"help": "child x", "metavar": "X"},
+                },
             ),
             types.OptionInfo(
                 parent_cls=GrandChild,
@@ -462,6 +503,11 @@ class TestDataclasses:
                 default=None,
                 has_no_default=False,
                 default_is_factory=False,
+                metadata={
+                    "argparse": {"help": "grand child x", "metavar": "GX"},
+                    "click": {"help": "grand child x", "metavar": "GX"},
+                    "help": "grand child x",
+                },
             ),
             types.OptionInfo(
                 parent_cls=Parent,
@@ -614,10 +660,26 @@ class TestPydantic:
         from pydantic_core._pydantic_core import PydanticUndefined
 
         class GrandChild(pydantic.BaseModel):
-            x: Optional[int] = None
+            x: Optional[int] = pydantic.Field(
+                default=None,
+                description="grand child x",
+                json_schema_extra={
+                    "typed-settings": {
+                        "argparse": {"metavar": "GX"},
+                        "click": {"metavar": "GX"},
+                    },
+                },
+            )
 
         class Child(pydantic.BaseModel):
-            x: "float"  # Test resolving types
+            x: "float" = pydantic.Field(  # Test resolving types
+                json_schema_extra={
+                    "typed-settings": {
+                        "help": "child x",
+                        "click": {"metavar": "X"},
+                    },
+                }
+            )
             y: GrandChild
 
         class Parent(pydantic.BaseModel):
@@ -642,6 +704,11 @@ class TestPydantic:
                 default=PydanticUndefined,
                 has_no_default=True,
                 default_is_factory=False,
+                metadata={
+                    "argparse": {"help": "child x"},
+                    "click": {"help": "child x", "metavar": "X"},
+                    "help": "child x",
+                },
             ),
             types.OptionInfo(
                 parent_cls=GrandChild,
@@ -650,6 +717,10 @@ class TestPydantic:
                 default=None,
                 has_no_default=False,
                 default_is_factory=False,
+                metadata={
+                    "argparse": {"help": "grand child x", "metavar": "GX"},
+                    "click": {"help": "grand child x", "metavar": "GX"},
+                },
             ),
             types.OptionInfo(
                 parent_cls=Parent,